moved phoenix.js to separate folder, fixed styles
